% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{hpa_download}
\alias{hpa_download}
\title{Download datasets}
\usage{
hpa_download(download_list = "histology")
}
\arguments{
\item{download_list}{A vector or string indicate which datasets to download. Possible value:
\itemize{
  \item \code{'Normal tissue'}
  \item \code{'Pathology'}
  \item \code{'Subcellular location'}
  \item \code{'RNA tissue'}
  \item \code{'RNA cell line'}
  \item \code{'RNA transcript tissue'}
  \item \code{'RNA transcript cell line'}
  \item \code{'all'}: download everything
  \item \code{'histology'}: same as \code{c('Normal tissue', 'Pathology', 'Subcellular location')}
  \item \code{'rna'}: same as \code{c('RNA tissue', 'RNA cell line')}
  \item \code{'isoform'}: same as \code{c('RNA transcript tissue', 'RNA transcript cell line')}
}
See \url{https://www.proteinatlas.org/about/download} for more information.}
}
\value{
This function will return a list of tibbles corresponding to requested datasets.
}
\description{
Download the latest version of HPA datasets and import them in R. It is
recommended to only download the datasets you need, as some of them may be
very big.
}
\examples{
  \dontrun{
  ## download rna expression data
  rna_data <- hpa_download('rna')
  
  ## download normal tissue and subcellular location data
  data <- hpa_download(c('Normal tissue', 'Subcellular location'))
  }
     
}
