% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpa_histology_data.R
\docType{data}
\name{hpa_downloaded_histology_v18}
\alias{hpa_downloaded_histology_v18}
\title{HPA histology dataset version 18}
\format{A list of 3 tibbles
\describe{
    \item{normal_tissue}{Normal tissue IHC data}
    \item{pathology}{Cancer IHC data}
    \item{subcellular_location}{Subcellular location IF data}
}}
\usage{
hpa_downloaded_histology_v18
}
\description{
Dataset downloaded with \code{hpaDownload('histology', version = 'v18')}.
This dataset is kept for the sake of backward compability. Please use
`hpa_histology_data` for the most updated built-in dataset.
}
\details{
Links to original data:
\itemize{
    \item {https://v18.proteinatlas.org/download/normal_tissue.tsv.zip}
    \item {https://v18.proteinatlas.org/download/pathology.tsv.zip}
    \item {https://v18.proteinatlas.org/download/subcellular_location.tsv.zip}
}
}
\examples{
  # load data
  data("hpa_downloaded_histology_v18")
  
  # access data frames
  normal_tissue_data <- hpa_downloaded_histology_v18$normal_tissue
  cancer_data <- hpa_downloaded_histology_v18$pathology
  subcell_location_data <- hpa_downloaded_histology_v18$subcellular_location
}
\seealso{
\code{\link{hpaDownload}}
\code{\link{hpa_histology_data}}
}
\keyword{datasets}
