% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpa_downloaded_rna_v18}
\alias{hpa_downloaded_rna_v18}
\title{HPA RNA dataset}
\format{A list of 2 tibbles
\describe{
    \item{rna_tissue}{Normal tissue RNAseq data}
    \item{rna_cell_line}{Cell line RNA seq data}
    }}
\usage{
hpa_downloaded_rna_v18
}
\description{
Dataset downloaded with \code{hpa_download('rna')}
}
\keyword{datasets}
