% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{hpa_get_xml}
\alias{hpa_get_xml}
\title{Download and import xml file}
\usage{
hpa_get_xml(target_ensembl_id)
}
\arguments{
\item{target_ensembl_id}{A string of one ensembl ID, start with ESNG. For
example \code{'ENSG00000134057'}}
}
\value{
This function return an object of class \code{"xml_document"
  "xml_node"}. See documentations for package \code{xml2} for more
  information.
}
\description{
Download and import individual xml file for a specified protein. This
function calls \code{xml2::read_xml()} under the hood.
}
\examples{
  \dontrun{
  CCNB1_xml <- hpa_get_xml('ENSG00000134057')
  }

}
