% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{hpa_list_param}
\alias{hpa_list_param}
\title{List available data}
\usage{
hpa_list_param(data)
}
\arguments{
\item{data}{Input the list object generated by \code{hpa_download()}}
}
\value{
The output of this function is a list of vectors.
}
\description{
List available variables in downloaded data that can be used as parameters to
subset the data via \code{hpa_subset()}. This function work with the data
object generated by \code{hpa_download()}.
}
\examples{
  data("hpa_downloaded_histology_v18")
  params <- hpa_list_param(data = hpa_downloaded_histology_v18)
  params$normal_tissue
  \dontrun{
  downloaded_data <- hpa_download(download_list = 'histology')
  params <- hpa_list_param(data = downloaded_data)
  params$normal_tissue
  }

}
