% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{hpa_vis_pathology}
\alias{hpa_vis_pathology}
\title{Visualize pathology data}
\usage{
hpa_vis_pathology(data, target_gene, target_cancer = NULL,
  color = c("#ffffb2", "#fecc5c", "#fd8d3c", "#e31a1c"),
  custom_theme = FALSE)
}
\arguments{
\item{data}{Input the list object generated by \code{hpa_download()} or
\code{hpa_subset()}. Require the \code{pathology} dataset.}

\item{target_gene}{Vector of strings of HGNC gene symbols.}

\item{target_cancer}{Vector of strings of normal tissues. The function will
plot all available cancer by default.}

\item{color}{Vector of 4 colors used to depict different expression levels.}

\item{custom_theme}{Logical argument. If \code{TRUE}, the function will
return a barebone ggplot2 plot to be customized further.}
}
\value{
This function will return a ggplot2 plot object, which can be further
  modified if desirable.
}
\description{
Visualize the expression of genes of interest in each cancer.
}
\examples{
  data("hpa_downloaded_histology_v18")
  gene_list <- c('TP53', 'EGFR', 'CD44', 'PTEN', 'IDH1', 'IDH2', 'CYCS')
  cancer_list <- c('breast cancer', 'glioma', 'melanoma')

  ## A typical function call
  hpa_vis_pathology(data = hpa_downloaded_histology_v18,
                 target_gene = gene_list)

  \dontrun{
  ## With more customization
  color_list <- c('yellow', 'red', 'green', 'blue')
  plot <- hpa_vis_pathology(data = hpa_downloaded_histology_v18,
                            target_gene = gene_list,
                            target_cancer = cancer_list,
                            color = color_list,
                            custom_theme = TRUE)
  plot + theme_minimal()
  }


}
