% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{hpa_vis_subcell}
\alias{hpa_vis_subcell}
\title{Visualize subcellular location data}
\usage{
hpa_vis_subcell(data, target_gene, color = c("white", "black"),
  custom_theme = FALSE)
}
\arguments{
\item{data}{Input the list object generated by \code{hpa_download()} or
\code{hpa_subset()}. Require the \code{subcellular_location} dataset.}

\item{target_gene}{Vector of strings of HGNC gene symbols.}

\item{color}{Vector of 2 colors used to depict if the protein expresses in a
location or not.}

\item{custom_theme}{Logical argument. If \code{TRUE}, the function will
return a barebone ggplot2 plot to be customized further.}
}
\value{
This function will return a ggplot2 plot object, which can be further
  modified if desirable.
}
\description{
Visualize the the confirmed subcellular locations of genes of interest.
}
\examples{
  \dontrun{
  downloaded_data <- hpa_download(download_list = 'Subcellular location')
  gene_list <- c('TP53', 'EGFR', 'CD44', 'PTEN', 'IDH1', 'IDH2', 'CYCS')
  
  ## A typical function call
  hpa_vis_subcell(data = downloaded_data,
                  target_gene = gene_list)
  
  ## With more customization
  color_list <- c('white', 'red')
  plot <- hpa_vis_subcell(data = downloaded_data,
                          target_gene = gene_list,
                          color = color_list,
                          custom_theme = TRUE)
  plot + theme_minimal()
  }


}
