% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{hpa_xml_protein_classes}
\alias{hpa_xml_protein_classes}
\title{Extract protein classes}
\usage{
hpa_xml_protein_classes(imported_xml)
}
\arguments{
\item{imported_xml}{Input an xml document object resulted from a \code{hpa_get_xml()} call.}
}
\value{
This function return a tibble of 4 columns.
}
\description{
Extract protein class information from imported xml document resulted from \code{hpa_get_xml()}.
}
\examples{
  \dontrun{
  CCNB1_xml <- hpa_get_xml('ENSG00000134057')
  hpa_xml_protein_classes(CCNB1_xml)
  }

}
