% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{hpa_xml_tissue_expr}
\alias{hpa_xml_tissue_expr}
\title{Extract tissue expression and download images}
\usage{
hpa_xml_tissue_expr(imported_xml, download_img = FALSE)
}
\arguments{
\item{imported_xml}{Input an xml document object resulted from a
\code{hpa_get_xml()} call.}

\item{download_img}{Logical argument. The function will download all image
from the extracted urls into the working folder.}
}
\value{
This function return a list consists of a summary string and a tibble of 2 columns.
}
\description{
Extract tissue expression information and url to download images from
imported xml document resulted from \code{hpa_get_xml()}.
}
\examples{
  \dontrun{
  CCNB1_xml <- hpa_get_xml('ENSG00000134057')
  hpa_xml_tissue_expr(CCNB1_xml)
  }
  
}
