% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{hpaListParam}
\alias{hpaListParam}
\title{List available data}
\usage{
hpaListParam(data)
}
\arguments{
\item{data}{Input the list object generated by \code{hpaDownload()}}
}
\value{
The output of this function is a list of vectors.
}
\description{
List available variables in downloaded data that can be used as parameters to
subset the data via \code{hpaSubset()}. This function work with the data
object generated by \code{hpaDownload()}.
}
\examples{
  downloadedData <- hpaDownload(downloadList='all', version='example')
  params <- hpaListParam(data=downloadedData)
  params$normal_tissue

}
