% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{hpaSubset}
\alias{hpaSubset}
\title{Subset downloaded data}
\usage{
hpaSubset(data, targetGene = NULL, targetTissue = NULL,
  targetCellType = NULL, targetCancer = NULL, targetCellLine = NULL)
}
\arguments{
\item{data}{Input the list object generated by \code{hpaDownload()} or
\code{hpaSubset()}}

\item{targetGene}{Vector of strings of HGNC gene symbols. Will be used to
subset every dataset in the list object.}

\item{targetTissue}{Vector of strings of normal tissues. Will be used to
subset the \code{normal_tissue} and \code{rna_tissue} dataset.}

\item{targetCellType}{Vector of strings of normal cell types. Will be used
to subset the \code{normal_tissue} dataset.}

\item{targetCancer}{Vector of strings of cancer types. Will be used to
subset the \code{pathology} dataset.}

\item{targetCellLine}{Vector of strings of cell lines. Will be used to
subset the \code{rna_cell_line} dataset.}
}
\value{
This function will return a list of tibbles.
}
\description{
Subset data by gene name, tissue, cell type, cancer and/or cell line. The
input is the list object generated by \code{hpaDownload()} or as the output
of another \code{hpaSubset()}. Use \code{hpaListParam()} to see the list
of available parameters for a specific list object. Will not work on isoform
data.
}
\examples{
  downloaded_data <- hpaDownload(downloadList = 'all', version = 'example')
  gene_list <- c('TP53', 'EGFR')
  tissue_list <- c('breast', 'cerebellum', 'skin 1')
  cancer_list <- c('breast cancer', 'glioma', 'melanoma')

  subset_data <- hpaSubset(data = downloaded_data,
                            targetGene = gene_list,
                            targetTissue = tissue_list,
                            targetCancer = cancer_list)
  \dontrun{
  further_subset_data <- hpaSubset(data = subset_data,
                                    targetGene = 'TP53')
  }

}
