% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{hpaXmlRNAExpr}
\alias{hpaXmlRNAExpr}
\title{Extract RNA expression details}
\usage{
hpaXmlRNAExpr(importedXml, section = "tissue", child = "level")
}
\arguments{
\item{importedXml}{Input an xml document object resulted from a
\code{hpaXmlGet()} call.}

\item{section}{Select the section to be extracted from xml including:
tissue, cellLine, or bloodCell.}

\item{child}{Select the multiple child nodes to extract from section
including: level or RNASample.}
}
\value{
This function returns a data frame of listed RNA expressions. 
  Each data frame contains information about all entries' RNA expression.
  Due to the variation in amount of information different parameters
  return different data frames by section (tissue, cellLine, or 
  bloodCell) with either level or RNASample information.
}
\description{
Extract RNA expression information for each sample and url to download
images from imported xml document resulted from \code{hpaXmlGet()}.
}
\examples{
\dontrun{
  GCH1xml <- hpaXmlGet('ENSG00000131979')
  rna_df <- hpaXmlRNAExpr(GCH1xml)
  rna_df <- hpaXmlRNAExpr(GCH1xml, section="cellLine", child="level")
  rna_df <- hpaXmlRNAExpr(GCH1xml, section="bloodCell", child="RNASample")
}

}
\seealso{
Other xml functions: 
\code{\link{hpaXmlAntibody}()},
\code{\link{hpaXmlGet}()},
\code{\link{hpaXmlProtClass}()},
\code{\link{hpaXmlTissueExprSum}()},
\code{\link{hpaXmlTissueExpr}()},
\code{\link{hpaXml}()}
}
\concept{xml functions}
