% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpa_downloaded_histology_v18.R
\docType{data}
\name{hpa_downloaded_histology_v18}
\alias{hpa_downloaded_histology_v18}
\title{HPA histology dataset}
\format{A list of 3 tibbles
\describe{
    \item{normal_tissue}{Normal tissue IHC data}
    \item{pathology}{Cancer IHC data}
    \item{subcellular_location}{Subcellular location IF data}
    }}
\usage{
hpa_downloaded_histology_v18
}
\description{
Dataset downloaded with \code{hpaDownload('histology', version = 'v18')}
}
\keyword{datasets}
