% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{hpaDownload}
\alias{hpaDownload}
\title{Download datasets}
\usage{
hpaDownload(downloadList = "histology", version = "latest")
}
\arguments{
\item{downloadList}{A vector or string indicate which datasets to download.
Possible value:
\itemize{
  \item \code{'Normal tissue'}
  \item \code{'Pathology'}
  \item \code{'Subcellular location'}
  \item \code{'RNA tissue'}
  \item \code{'RNA cell line'}
  \item \code{'RNA transcript tissue'}
  \item \code{'RNA transcript cell line'}
  \item \code{'all'}: download everything
  \item \code{'histology'}: same as \code{c('Normal tissue', 'Pathology',
  'Subcellular location')}
  \item \code{'rna'}: same as \code{c('RNA tissue', 'RNA cell line')}
  \item \code{'isoform'}: same as \code{c('RNA transcript tissue', 'RNA
  transcript cell line')}
}
See \url{https://www.proteinatlas.org/about/download} for more information.}

\item{version}{A string indicate which version to be downloaded. Possible
value:
\itemize{
  \item \code{'hpar'}: Load data from the 'hpar' package. Does not contain
  isoform data.
  \item \code{'latest'}: Download latest version. Require Internet
  connection.
  \item \code{'example'}: Load the example dataset from 'HPA
  analyze' ('hpa_downloaded_histology_v18'). Does not contain rna or
  isoform data.
  \item \code{'v?'} with '?' is a integer: Download a specific version of
  the dataset. For example: 'v18' download version 18. Currently support
  version 17 and above. Require Internet connection.
}}
}
\value{
This function will return a list of tibbles corresponding to
  requested datasets.
}
\description{
Download the latest version of HPA datasets and import them in R. It is
recommended to only download the datasets you need, as some of them may be
very big.
}
\examples{
  downloadedData <- hpaDownload(downloadList='all', version='example')
  summary(downloadedData)
  
}
\seealso{
Other downloadable datasets functions: \code{\link{hpaExport}},
  \code{\link{hpaSubset}}
}
\concept{downloadable datasets functions}
