% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{hpaDownload}
\alias{hpaDownload}
\title{Download datasets}
\usage{
hpaDownload(downloadList = "histology", version = "latest", tidyNames = TRUE)
}
\arguments{
\item{downloadList}{A vector or string indicate which datasets to download.
  Common values:
  \itemize{
    \item \code{'Normal tissue'}
    \item \code{'Pathology'}
    \item \code{'Subcellular location'}
    }
  For the full list of possible values for a specific version, set
  downloadList as NULL: \code{hpaDownload(downloadList = NULL, version =
  <version>)}

  You can also use the following shortcuts:
  \itemize{
    \item \code{'all'}: download everything
    \item \code{'histology'}: same as \code{c('Normal tissue', 'Pathology',
    'Subcellular location')}
    \item \code{'rna tissue'}: same as \code{c('RNA consensus tissue', 'RNA
    HPA tissue', 'RNA GTEx tissue', 'RNA FANTOM tissue')}
    \item \code{'rna cell type'}: same as \code{c('RNA single cell
    type', 'RNA single cell type tissue cluster')}
    \item \code{'rna brain region'}: same as \code{c('RNA GTEx brain region',
    'RNA FANTOM brain region', 'RNA pig brain region', 'RNA pig brain
    subregion sample', 'RNA mouse brain region', 'RNA mouse brain subregion
    sample', 'RNA Allen mouse brain region')}
    \item \code{'rna immune cell'}: same as \code{c('RNA HPA immune
    cell', 'RNA HPA immune cell sample', 'RNA Monaco immune cell', 'RNA
    Schmiedel immune cell')}
    \item \code{'rna blood cell'}: same as \code{c('RNA HPA blood
    cell', 'RNA HPA blood cell sample', 'RNA Monaco blood cell', 'RNA
    Schmiedel blood cell')}
    \item \code{'isoform'}: same as \code{c('RNA isoform tissue', 'RNA
    isoform GTEx retina', 'RNA isoform immune cells', 'RNA
    isoform cell line', 'RNA isoform pig brain', 'RNA isoform mouse
    brain')}
  }
  See \url{https://www.proteinatlas.org/about/download} for more information.}

\item{version}{A string indicate which version to be downloaded. Possible
value:
\itemize{
  \item \code{'latest'}: Download latest version. Certain legacy datasets
  will be downloaded with the highest version available. Require Internet
  connection. This is the default option.
  \item \code{'example'} or \code{'built-in'}: Load the built-in histology
  dataset from 'HPAanalyze' ('hpa_histology_data'). Do not require internet
  connection.
}}

\item{tidyNames}{TRUE or FALSE. Indicate if 'Normal tissue', 'Pathology',
'Subcellular location' should be renamed with tidy column names. Default is
TRUE for backward compatibility reasons.}
}
\value{
This function will return a list of tibbles corresponding to
  requested datasets.
}
\description{
Download the latest version of HPA datasets and import them in R. It is
recommended to only download the datasets you need, as some of them may be
very big.
}
\examples{
  histologyData <- hpaDownload(downloadList='histology', version='example')
  # tissueTranscriptData <- hpaDownload('RNA transcript tissue')

}
\seealso{
\code{\link{hpaDownload}} \code{\link{hpa_histology_data}}

Other downloadable datasets functions: 
\code{\link{hpaExport}()},
\code{\link{hpaSubset}()}
}
\concept{downloadable datasets functions}
