% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{hpaExport}
\alias{hpaExport}
\title{Export the subset data}
\usage{
hpaExport(data, file_name, file_type = "xlsx")
}
\arguments{
\item{data}{Input the list object generated by \code{hpaSubset()}}

\item{file_name}{A string indicate the desired output file name}

\item{file_type}{For compability of future development. Currently the only
option is \code{'xlsx'}}
}
\value{
This function will create an xlsx file which one individual
  spreadsheet for each datasets in the input list object.
}
\description{
Export the list object generated by \code{hpaSubset()} into xlsx format. Due
to the size of some HPA datasets, as well as the limitation of the output
format, exporting the full datasets generated by \code{hpaDownload()} is not
recommended.
}
\examples{
  data("hpa_downloaded_histology_v18")
  gene_list <- c('TP53', 'EGFR')
  tissue_list <- c('breast', 'cerebellum', 'skin 1')
  cancer_list <- c('breast cancer', 'glioma', 'melanoma')

  subset_data <- hpaSubset(data = hpa_downloaded_histology_v18,
                            target_gene = gene_list,
                            target_tissue = tissue_list,
                            target_cancer = cancer_list)
  hpaExport(data = subset_data,
             file_name = 'TP53_EGFR_in_tissue_cancer.xlsx',
             file_type = 'xlsx')

}
