% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{hpaExport}
\alias{hpaExport}
\title{Export the subset data}
\usage{
hpaExport(data, fileName, fileType = "xlsx")
}
\arguments{
\item{data}{Input the list object generated by \code{hpaSubset()}}

\item{fileName}{A string indicate the desired output file name}

\item{fileType}{For compability of future development. Currently the only
option is \code{'xlsx'}}
}
\value{
This function will create an xlsx file which one individual
  spreadsheet for each datasets in the input list object.
}
\description{
Export the list object generated by \code{hpaSubset()} into xlsx format. Due
to the size of some HPA datasets, as well as the limitation of the output
format, exporting the full datasets generated by \code{hpaDownload()} is not
recommended.
}
\examples{
  downloadedData <- hpaDownload(downloadList='all', version='example')
  geneList <- c('TP53', 'EGFR')
  tissueList <- c('breast', 'cerebellum', 'skin 1')
  cancerList <- c('breast cancer', 'glioma', 'melanoma')

  subsetData <- hpaSubset(data=downloadedData,
                           targetGene=geneList,
                           targetTissue=tissueList,
                           targetCancer=cancerList)
  hpaExport(data=subsetData,
            fileName='TP53_EGFR_in_tissue_cancer.xlsx',
            fileType='xlsx')

}
