% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{hpaListParam}
\alias{hpaListParam}
\title{List available data}
\usage{
hpaListParam(data)
}
\arguments{
\item{data}{Input the list object generated by \code{hpaDownload()}}
}
\value{
The output of this function is a list of vectors.
}
\description{
List available variables in downloaded data that can be used as parameters to
subset the data via \code{hpaSubset()}. This function work with the data
object generated by \code{hpaDownload()}.
}
\examples{
  data("hpa_downloaded_histology_v18")
  params <- hpaListParam(data = hpa_downloaded_histology_v18)
  params$normal_tissue
  \dontrun{
  downloaded_data <- hpaDownload(download_list = 'histology')
  params <- hpaListParam(data = downloaded_data)
  params$normal_tissue
  }

}
