% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{hpaSubset}
\alias{hpaSubset}
\title{Subset downloaded data}
\usage{
hpaSubset(data, target_gene = NULL, target_tissue = NULL,
  target_cell_type = NULL, target_cancer = NULL, target_cell_line = NULL)
}
\arguments{
\item{data}{Input the list object generated by \code{hpaDownload()} or
\code{hpaSubset()}}

\item{target_gene}{Vector of strings of HGNC gene symbols. Will be used to
subset every dataset in the list object.}

\item{target_tissue}{Vector of strings of normal tissues. Will be used to
subset the \code{normal_tissue} and \code{rna_tissue} dataset.}

\item{target_cell_type}{Vector of strings of normal cell types. Will be used
to subset the \code{normal_tissue} dataset.}

\item{target_cancer}{Vector of strings of cancer types. Will be used to
subset the \code{pathology} dataset.}

\item{target_cell_line}{Vector of strings of cell lines. Will be used to
subset the \code{rna_cell_line} dataset.}
}
\value{
This function will return a list of tibbles.
}
\description{
Subset data by gene name, tissue, cell type, cancer and/or cell line. The
input is the list object generated by \code{hpaDownload()} or as the output
of another \code{hpaSubset()}. Use \code{hpaListParam()} to see the list
of available parameters for a specific list object. Will not work on isoform
data.
}
\examples{
  data("hpa_downloaded_histology_v18")
  gene_list <- c('TP53', 'EGFR')
  tissue_list <- c('breast', 'cerebellum', 'skin 1')
  cancer_list <- c('breast cancer', 'glioma', 'melanoma')

  subset_data <- hpaSubset(data = hpa_downloaded_histology_v18,
                            target_gene = gene_list,
                            target_tissue = tissue_list,
                            target_cancer = cancer_list)
  \dontrun{
  further_subset_data <- hpaSubset(data = subset_data,
                                    target_gene = 'TP53')
  }

}
