% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{hpaVisPatho}
\alias{hpaVisPatho}
\title{Visualize pathology data}
\usage{
hpaVisPatho(data = NULL, targetGene = NULL, targetCancer = NULL,
  color = c("#ffffb2", "#fecc5c", "#fd8d3c", "#e31a1c"),
  customTheme = FALSE)
}
\arguments{
\item{data}{Input the list object generated by \code{hpa_download()} or
\code{hpa_subset()}. Require the \code{pathology} dataset.}

\item{targetGene}{Vector of strings of HGNC gene symbols.}

\item{targetCancer}{Vector of strings of normal tissues. The function will
plot all available cancer by default.}

\item{color}{Vector of 4 colors used to depict different expression levels.}

\item{customTheme}{Logical argument. If \code{TRUE}, the function will return
a barebone ggplot2 plot to be customized further.}
}
\value{
This function will return a ggplot2 plot object, which can be further
  modified if desirable. The pathology data is visualized as multiple bar
  graphs, one for each type of cancer. For each bar graph, x axis contains
  the inquired protein and y axis contains the proportion of patients.
}
\description{
Visualize the expression of genes of interest in each cancer.
}
\examples{
  data("hpa_downloaded_histology_v18")
  geneList <- c('TP53', 'EGFR', 'CD44', 'PTEN', 'IDH1', 'IDH2', 'CYCS')
  cancerList <- c('breast cancer', 'glioma', 'melanoma')

  ## A typical function call
  hpaVisPatho(data=hpa_downloaded_histology_v18,
                 targetGene=geneList)

}
\seealso{
Other visualization functions: \code{\link{hpaVisSubcell}},
  \code{\link{hpaVisTissue}}, \code{\link{hpaVis}}
}
\concept{visualization functions}
