% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{hpaVisTissue}
\alias{hpaVisTissue}
\title{Visualize tissue data}
\usage{
hpaVisTissue(data, target_gene, target_tissue, target_cell_type = NULL,
  color = c("#ffffb2", "#fecc5c", "#fd8d3c", "#e31a1c"),
  custom_theme = FALSE)
}
\arguments{
\item{data}{Input the list object generated by \code{hpa_download()} or
\code{hpa_subset()}. Require the \code{normal_tissue} dataset.}

\item{target_gene}{Vector of strings of HGNC gene symbols.}

\item{target_tissue}{Vector of strings of normal tissues.}

\item{target_cell_type}{Vector of strings of normal cell types.}

\item{color}{Vector of 4 colors used to depict different expression levels.}

\item{custom_theme}{Logical argument. If \code{TRUE}, the function will
return a barebone ggplot2 plot to be customized further.}
}
\value{
This function will return a ggplot2 plot object, which can be further
  modified if desirable.
}
\description{
Visualize the expression of protein of interest in each target tissue by cell
types.
}
\examples{
  data("hpa_downloaded_histology_v18")
  gene_list <- c('TP53', 'EGFR', 'CD44', 'PTEN', 'IDH1', 'IDH2', 'CYCS')
  tissue_list <- c('breast', 'cerebellum', 'skin 1')

  ## A typical function call   
  hpaVisTissue(data = hpa_downloaded_histology_v18,
                 target_gene = gene_list,
                 target_tissue = tissue_list)
                 
  \dontrun{               
  ## With more customization
  color_list <- c('white', 'red', 'green', 'blue')
  plot <- hpaVisTissue(data = hpa_downloaded_histology_v18,
                         target_gene = gene_list,
                         target_tissue = tissue_list,
                         color = color_list,
                         custom_theme = TRUE)
  plot + theme_minimal()
  }

}
