% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{hpaXmlAntibody}
\alias{hpaXmlAntibody}
\title{Extract antibody information}
\usage{
hpaXmlAntibody(importedXml)
}
\arguments{
\item{importedXml}{Input an xml document object resulted from a
\code{hpaXmlGet()} call.}
}
\value{
This function returns a tibble of 4 columns.
}
\description{
Extract information about the antibodies used for a specific protein.
}
\examples{
  CCNB1xml <- hpaXmlGet('ENSG00000134057')
  hpaXmlAntibody(CCNB1xml)
  
}
