% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{hpaXmlAntibody}
\alias{hpaXmlAntibody}
\title{Extract antibody information}
\usage{
hpaXmlAntibody(importedXml)
}
\arguments{
\item{importedXml}{Input an xml document object resulted from a
\code{hpaXmlGet()} call.}
}
\value{
This function returns a tibble of 4 columns, containing information
  about the antibodies used in the project for the inquired protein: id,
  releaseDate, releaseVersion, and RRID.
}
\description{
Extract information about the antibodies used for a specific protein.
}
\examples{
  GCH1xml <- hpaXmlGet('ENSG00000131979')
  hpaXmlAntibody(GCH1xml)

}
