% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{hpaXmlGet}
\alias{hpaXmlGet}
\title{Download and import xml file}
\usage{
hpaXmlGet(targetEnsemblId, version = NULL)
}
\arguments{
\item{targetEnsemblId}{A string of one ensembl ID, start with ESNG. For
example \code{'ENSG00000134057'}}

\item{version}{A string, indicate which version of data to be downloaded. The
latest version is downloaded by default. (This option is currently under
development.)}
}
\value{
This function return an object of class \code{"xml_document"
  "xml_node"}. See documentations for package \code{xml2} for more
  information.
}
\description{
Download and import individual xml file for a specified protein. This
function calls \code{xml2::read_xml()} under the hood.
}
\examples{
  print('Please run the example below in your console.')
  \dontrun{
  CCNB1_xml <- hpaXmlGet('ENSG00000134057')
  }

}
