% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{hpaXmlGet}
\alias{hpaXmlGet}
\title{Download and import xml file}
\usage{
hpaXmlGet(targetEnsemblId, version = "latest")
}
\arguments{
\item{targetEnsemblId}{A string of one ensembl ID, start with ESNG. For
example \code{'ENSG00000134057'}}

\item{version}{A string indicate which version to be downloaded. Possible
value:
\itemize{
  \item \code{'latest'}: Download latest version. 
  \item \code{'v?'} with '?' is a integer: Download a specific version of
  the dataset. For example: 'v18' download version 18. Currently support
  version 13 and above.
}}
}
\value{
This function return an object of class \code{"xml_document"
  "xml_node"}. See documentations for package \code{xml2} for more
  information.
}
\description{
Download and import individual xml file for a specified protein. This
function calls \code{xml2::read_xml()} under the hood.
}
\examples{
  print('Please run the example below in your console.')
  \dontrun{
  CCNB1xml <- hpaXmlGet('ENSG00000134057')
  }

}
