% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{hpaXmlProtClass}
\alias{hpaXmlProtClass}
\title{Extract protein classes}
\usage{
hpaXmlProtClass(importedXml)
}
\arguments{
\item{importedXml}{Input an xml document object resulted from a
\code{hpaXmlGet()} call.}
}
\value{
This function return a tibble of 4 columns.
}
\description{
Extract protein class information from imported xml document resulted from
\code{hpaXmlGet()}.
}
\examples{
  GCH1xml <- hpaXmlGet('ENSG00000131979')
  hpaXmlProtClass(GCH1xml)

}
\seealso{
Other xml functions: \code{\link{hpaXmlAntibody}},
  \code{\link{hpaXmlGet}},
  \code{\link{hpaXmlTissueExprSum}},
  \code{\link{hpaXmlTissueExpr}}, \code{\link{hpaXml}}
}
\concept{xml functions}
