% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{hpaXmlProtClass}
\alias{hpaXmlProtClass}
\title{Extract protein classes}
\usage{
hpaXmlProtClass(imported_xml)
}
\arguments{
\item{imported_xml}{Input an xml document object resulted from a \code{hpaXmlGet()} call.}
}
\value{
This function return a tibble of 4 columns.
}
\description{
Extract protein class information from imported xml document resulted from \code{hpaXmlGet()}.
}
\examples{
  print('Please run the example below in your console.')
  \dontrun{
  CCNB1_xml <- hpaXmlGet('ENSG00000134057')
  hpaXmlProtClass(CCNB1_xml)
  }

}
