% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{hpaXmlTissueExpr}
\alias{hpaXmlTissueExpr}
\title{Extract tissue expression details}
\usage{
hpaXmlTissueExpr(importedXml)
}
\arguments{
\item{importedXml}{Input an xml document object resulted from a
\code{hpaXmlGet()} call.}
}
\value{
This function returns a list of tibbles, each for an antibody.
}
\description{
Extract tissue expression information for each sample and url to download
images from imported xml document resulted from \code{hpaXmlGet()}.
}
\examples{
  print('Please run the example below in your console.')
  \dontrun{
  CCNB1xml <- hpaXmlGet('ENSG00000134057')
  hpaXmlTissueExpr(CCNB1xml)
  }

}
