% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{hpaXmlTissueExpr}
\alias{hpaXmlTissueExpr}
\title{Extract tissue expression and download images}
\usage{
hpaXmlTissueExpr(imported_xml, download_img = FALSE)
}
\arguments{
\item{imported_xml}{Input an xml document object resulted from a
\code{hpaXmlGet()} call.}

\item{download_img}{Logical argument. The function will download all image
from the extracted urls into the working folder.}
}
\value{
This function return a list consists of a summary string and a tibble of 2 columns.
}
\description{
Extract tissue expression information and url to download images from
imported xml document resulted from \code{hpaXmlGet()}.
}
\examples{
  print('Please run the example below in your console.')
  \dontrun{
  CCNB1_xml <- hpaXmlGet('ENSG00000134057')
  hpaXmlTissueExpr(CCNB1_xml)
  }
  
}
